+++ CHANGELOG.md
@@ -70,6 +70,11 @@
   directly.  
   [Matt Taube](https://github.com/mtaube)
   [#715](https://github.com/realm/SwiftLint/pull/715)
+  
+* Add `ImplicitGetterRule` to warn against using `get` on computed read-only
+  properties.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+  [#57](https://github.com/realm/SwiftLint/issues/57)
 
 ##### Bug Fixes